refactor: move logging and workpoolOptions to the shared src root
Both are plain validators/helpers used by the component and the client alike,
so having them under component/ meant the client reached into the component to
get them. Neither defines any Convex functions, so the only codegen effect is
dropping two type-only imports from the component's generated api.ts.
