+++ src/component/_generated/api.ts
@@ -11,13 +11,11 @@
 import type * as actionRunner from "../actionRunner.js";
 import type * as event from "../event.js";
 import type * as journal from "../journal.js";
-import type * as logging from "../logging.js";
 import type * as model from "../model.js";
 import type * as oversizedValues from "../oversizedValues.js";
 import type * as pool from "../pool.js";
 import type * as utils from "../utils.js";
 import type * as workflow from "../workflow.js";
-import type * as workpoolOptions from "../workpoolOptions.js";
 
 import type {
   ApiFromModules,
@@ -30,13 +28,11 @@ const fullApi: ApiFromModules<{
   actionRunner: typeof actionRunner;
   event: typeof event;
   journal: typeof journal;
-  logging: typeof logging;
   model: typeof model;
   oversizedValues: typeof oversizedValues;
   pool: typeof pool;
   utils: typeof utils;
   workflow: typeof workflow;
-  workpoolOptions: typeof workpoolOptions;
 }> = anyApi as any;
 
 /**
