+++ src/component/utils.ts
@@ -1,6 +1,6 @@
 import { v } from "convex/values";
 import { internalMutation, type QueryCtx } from "./_generated/server.js";
-import { createLogger, DEFAULT_LOG_LEVEL, logLevel } from "./logging.js";
+import { createLogger, DEFAULT_LOG_LEVEL, logLevel } from "../logging.js";
 
 export async function getDefaultLogger(ctx: QueryCtx) {
   const config = await ctx.db.query("config").first();
